+++ Vedeath/Assembly-CSharp.csproj
@@ -51,9 +51,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-     <Compile Include="Assets\Enemy_Instancing.cs" />
      <Compile Include="Assets\Scripts\Bullet.cs" />
      <Compile Include="Assets\Scripts\Enemy\Enemy_Movement.cs" />
+     <Compile Include="Assets\Scripts\Enemy_Instancing.cs" />
      <Compile Include="Assets\Scripts\PlayerController.cs" />
      <Compile Include="Assets\Scripts\Shooting.cs" />
  <Reference Include="UnityEngine.UI">
