--- Vedeath/Assets/Enemy_Instancing.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 1443b1fa084f24ba695d7c9c135bb02f
-timeCreated: 1527823354
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
